Custom Markdown component done and added.
