+++ frontend/app/components/CustomMarkdown/types.ts
@@ -1,0 +1,6 @@
+export interface CustomMarkdownProps {
+  content: string;
+  backgroundColor: string;
+  imageHeight?: string | number;
+  imageWidth?: string | number;
+}
